server/requirements.txt
APScheduler==3.6.1
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dnspython==1.16.0
eventlet==0.25.1
Flask==1.1.1
Flask-SocketIO==4.2.1
greenlet==0.4.15
idna==2.8
importlib-metadata==0.22
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
monotonic==1.5
more-itertools==7.2.0
packaging==19.1
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.2
pytest==5.1.2
python-engineio==3.9.3
python-socketio==4.3.1
pytz==2019.2
requests==2.22.0
six==1.12.0
tzlocal==2.0.0
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.6
zipp==0.6.0
